refactor: rename explosion.v2 package to explosion.domain
