+++ src/main/java/net/qsef1256/dacobot/game/explosion/command/InventoryCommand.java
@@ -10,8 +10,8 @@ import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 import net.dv8tion.jda.api.interactions.commands.build.SubcommandGroupData;
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.qsef1256.dacobot.core.ui.command.DacoCommand;
-import net.qsef1256.dacobot.game.explosion.v2.inventory.InventoryController;
-import net.qsef1256.dacobot.game.explosion.v2.inventory.InventoryService;
+import net.qsef1256.dacobot.game.explosion.domain.inventory.InventoryController;
+import net.qsef1256.dacobot.game.explosion.domain.inventory.InventoryService;
 import net.qsef1256.dacobot.ui.DiaEmbed;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Component;
