+++ src/main/java/net/qsef1256/dacobot/game/explosion/domain/cash/CashRepository.java
@@ -1,0 +1,7 @@
+package net.qsef1256.dacobot.game.explosion.domain.cash;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CashRepository extends JpaRepository<CashEntity, Long> {
+
+}
