+++ src/main/java/net/qsef1256/dacobot/game/explosion/domain/inventory/InventoryRepository.java
@@ -1,0 +1,7 @@
+package net.qsef1256.dacobot.game.explosion.domain.inventory;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+
+}
