+++ src/main/java/net/qsef1256/dacobot/game/explosion/domain/itemtype/ItemTypeCommand.java
@@ -1,0 +1,19 @@
+package net.qsef1256.dacobot.game.explosion.domain.itemtype;
+
+import com.jagrosh.jdautilities.command.SlashCommandEvent;
+import net.qsef1256.dacobot.core.ui.command.DacoCommand;
+import org.jetbrains.annotations.NotNull;
+
+public class ItemTypeCommand extends DacoCommand {
+
+    public ItemTypeCommand() {
+        name = "아이템";
+        help = "아이템 타입 관리 명령어";
+    }
+
+    @Override
+    protected void runCommand(@NotNull SlashCommandEvent event) {
+        // TODO: fill this
+    }
+
+}
