+++ src/main/java/net/qsef1256/dacobot/game/explosion/domain/itemtype/ItemTypeEntity.java
@@ -1,0 +1,38 @@
+package net.qsef1256.dacobot.game.explosion.domain.itemtype;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
+
+@Getter
+@Entity
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "explosion_item_type")
+public class ItemTypeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer itemId;
+
+    @Column(nullable = false)
+    private String itemName;
+
+    @Column(columnDefinition = "TEXT")
+    private String itemIcon = "";
+
+    @Column(columnDefinition = "TEXT")
+    private String description;
+
+    @Enumerated(EnumType.STRING)
+    private ItemRank itemRank;
+
+    @Column(nullable = false)
+    @ColumnDefault(value = "1")
+    private Integer maxAmount;
+
+}
