--- src/main/java/net/qsef1256/dacobot/game/explosion/v2/cash/CashRepository.java
@@ -1,7 +1,0 @@
-package net.qsef1256.dacobot.game.explosion.v2.cash;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CashRepository extends JpaRepository<CashEntity, Long> {
-
-}
