--- src/main/java/net/qsef1256/dacobot/game/explosion/v2/inventory/InventoryRepository.java
@@ -1,7 +1,0 @@
-package net.qsef1256.dacobot.game.explosion.v2.inventory;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
-
-}
