--- src/main/java/net/qsef1256/dacobot/game/explosion/v2/itemtype/ItemTypeCommand.java
@@ -1,19 +1,0 @@
-package net.qsef1256.dacobot.game.explosion.v2.itemtype;
-
-import com.jagrosh.jdautilities.command.SlashCommandEvent;
-import net.qsef1256.dacobot.core.ui.command.DacoCommand;
-import org.jetbrains.annotations.NotNull;
-
-public class ItemTypeCommand extends DacoCommand {
-
-    public ItemTypeCommand() {
-        name = "아이템";
-        help = "아이템 타입 관리 명령어";
-    }
-
-    @Override
-    protected void runCommand(@NotNull SlashCommandEvent event) {
-        // TODO: fill this
-    }
-
-}
