--- src/main/java/net/qsef1256/dacobot/game/explosion/v2/itemtype/ItemTypeMapper.java
@@ -1,13 +1,0 @@
-package net.qsef1256.dacobot.game.explosion.v2.itemtype;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.factory.Mappers;
-
-@Mapper
-public interface ItemTypeMapper {
-
-    ItemTypeMapper INSTANCE = Mappers.getMapper(ItemTypeMapper.class);
-
-    ItemTypeDto mapItemTypeDto(ItemTypeEntity itemType);
-
-}
