--- src/main/java/net/qsef1256/dacobot/game/explosion/v2/itemtype/ItemTypeRepository.java
@@ -1,9 +1,0 @@
-package net.qsef1256.dacobot.game.explosion.v2.itemtype;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemTypeRepository extends JpaRepository<ItemTypeEntity, Integer> {
-
-}
