+++ src/test/java/net/qsef1256/dacobot/game/explosion/domain/cash/CashEntityTest.java
@@ -1,0 +1,18 @@
+package net.qsef1256.dacobot.game.explosion.domain.cash;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+@DataJpaTest
+class CashEntityTest {
+
+    @Test
+    void testCash() {
+        assertDoesNotThrow(() -> {
+
+        });
+    }
+
+}
