--- src/test/java/net/qsef1256/dacobot/game/explosion/v2/cash/CashEntityTest.java
@@ -1,18 +1,0 @@
-package net.qsef1256.dacobot.game.explosion.v2.cash;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-
-@DataJpaTest
-class CashEntityTest {
-
-    @Test
-    void testCash() {
-        assertDoesNotThrow(() -> {
-
-        });
-    }
-
-}
